# docs/openapi.yaml
openapi: 3.1.0
info:
  title: x4-api
  version: 0.0.1
paths:
...
  /api/v1/wares:
    get:
      tags:
      - wares
      summary: List Wares
      operationId: list_wares_api_v1_wares_get
      parameters:
      - name: group
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Group
      - name: transport
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Transport
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WareSummary'
                title: Response List Wares Api V1 Wares Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/wares/{ware_id}:
    get:
      tags:
      - wares
      summary: Get Ware
      operationId: get_ware_api_v1_wares__ware_id__get
      parameters:
      - name: ware_id
        in: path
        required: true
        schema:
          type: string
          title: Ware Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WareDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/terraform/projects:
    get:
      tags:
      - terraform
      summary: List Terraform Projects
      operationId: list_terraform_projects_api_v1_terraform_projects_get
      parameters:
      - name: group_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TerraformProjectSummary'
                title: Response List Terraform Projects Api V1 Terraform Projects
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/terraform/projects/{project_id}:
    get:
      tags:
      - terraform
      summary: Get Terraform Project
      operationId: get_terraform_project_api_v1_terraform_projects__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerraformProjectDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/ware-groups/{group_id}:
    get:
      tags:
      - ware-groups
      summary: Get Ware Group
      operationId: get_ware_group_api_v1_ware_groups__group_id__get
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WareGroup'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProductionInput:
      properties:
        ware_id:
          type: string
          title: Ware Id
        amount:
          type: integer
          title: Amount
      additionalProperties: false
      type: object
      required:
      - ware_id
      - amount
      title: ProductionInput
    ProductionMethod:
      properties:
        method:
          type: string
          title: Method
        time_sec:
          type: number
          title: Time Sec
        amount:
          type: integer
          title: Amount
        workforce:
          anyOf:
          - type: integer
          - type: 'null'
          title: Workforce
        inputs:
          items:
            $ref: '#/components/schemas/ProductionInput'
          type: array
          title: Inputs
      additionalProperties: false
      type: object
      required:
      - method
      - time_sec
      - amount
      - workforce
      - inputs
      title: ProductionMethod
...
    TerraformDelivery:
      properties:
        ship_macro:
          type: string
          title: Ship Macro
        amount:
          type: integer
          title: Amount
        build_duration:
          anyOf:
          - type: integer
          - type: 'null'
          title: Build Duration
      additionalProperties: false
      type: object
      required:
      - ship_macro
      - amount
      - build_duration
      title: TerraformDelivery
    TerraformEffect:
      properties:
        stat:
          type: string
          title: Stat
        change:
          type: integer
          title: Change
        min_val:
          anyOf:
          - type: integer
          - type: 'null'
          title: Min Val
      additionalProperties: false
      type: object
      required:
      - stat
      - change
      - min_val
      title: TerraformEffect
    TerraformProjectDetail:
      properties:
        project_id:
          type: string
          title: Project Id
        group_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        duration:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration
        repeat_cooldown:
          anyOf:
          - type: integer
          - type: 'null'
          title: Repeat Cooldown
        resilient:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resilient
        resource_credits:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resource Credits
        effects:
          items:
            $ref: '#/components/schemas/TerraformEffect'
          type: array
          title: Effects
        resources:
          items:
            $ref: '#/components/schemas/TerraformResource'
          type: array
          title: Resources
        deliveries:
          items:
            $ref: '#/components/schemas/TerraformDelivery'
          type: array
          title: Deliveries
      additionalProperties: false
      type: object
      required:
      - project_id
      - group_id
      - name
      - description
      - duration
      - repeat_cooldown
      - resilient
      - resource_credits
      - effects
      - resources
      - deliveries
      title: TerraformProjectDetail
...
    TerraformProjectSummary:
      properties:
        project_id:
          type: string
          title: Project Id
        group_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        duration:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration
        repeat_cooldown:
          anyOf:
          - type: integer
          - type: 'null'
          title: Repeat Cooldown
        resilient:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resilient
        resource_credits:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resource Credits
      additionalProperties: false
      type: object
      required:
      - project_id
      - group_id
      - name
      - description
      - duration
      - repeat_cooldown
      - resilient
      - resource_credits
      title: TerraformProjectSummary
    TerraformResource:
      properties:
        ware_id:
          type: string
          title: Ware Id
        amount:
          type: integer
          title: Amount
      additionalProperties: false
      type: object
      required:
      - ware_id
      - amount
      title: TerraformResource
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WareDetail:
      properties:
        ware_id:
          type: string
          title: Ware Id
        name:
          type: string
          title: Name
        group_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Id
        transport:
          anyOf:
          - type: string
          - type: 'null'
          title: Transport
        volume:
          type: number
          title: Volume
        price_avg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price Avg
        tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
        price_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price Min
        price_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price Max
        storage_class:
          anyOf:
          - type: string
          - type: 'null'
          title: Storage Class
        restriction_licence:
          anyOf:
          - type: string
          - type: 'null'
          title: Restriction Licence
        use_threshold:
          anyOf:
          - type: number
          - type: 'null'
          title: Use Threshold
        owners:
          items:
            type: string
          type: array
          title: Owners
        illegal_factions:
          items:
            type: string
          type: array
          title: Illegal Factions
        production:
          items:
            $ref: '#/components/schemas/ProductionMethod'
          type: array
          title: Production
      additionalProperties: false
      type: object
      required:
      - ware_id
      - name
      - group_id
      - transport
      - volume
      - price_avg
      - tags
      - icon_url
      - price_min
      - price_max
      - storage_class
      - owners
      - illegal_factions
      - production
      title: WareDetail
    WareGroup:
      properties:
        group_id:
          type: string
          title: Group Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
        factory_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Factory Name
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        factory_map_icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Factory Map Icon
        factory_hud_icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Factory Hud Icon
        tier:
          anyOf:
          - type: integer
          - type: 'null'
          title: Tier
        priority:
          anyOf:
          - type: integer
          - type: 'null'
          title: Priority
      additionalProperties: false
      type: object
      required:
      - group_id
      - name
      - tags
      - factory_name
      - icon
      - factory_map_icon
      - factory_hud_icon
      - tier
      - priority
      title: WareGroup
    WareSummary:
      properties:
        ware_id:
          type: string
          title: Ware Id
        name:
          type: string
          title: Name
        group_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Id
        transport:
          anyOf:
          - type: string
          - type: 'null'
          title: Transport
        volume:
          type: number
          title: Volume
        price_avg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price Avg
        tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
      additionalProperties: false
      type: object
      required:
      - ware_id
      - name
      - group_id
      - transport
      - volume
      - price_avg
      - tags
      - icon_url
      title: WareSummary
...